Add svg download functionality + move downloadPNG to Helper
